wip: simplify and standardise style options
